Dev: review points addressed
