--- packages/arb-token-bridge-ui/src/token-bridge-sdk/core/approveNativeCurrency.ts
@@ -1,7 +1,0 @@
-import { Provider } from '@ethersproject/providers'
-import { Signer } from 'ethers'
-
-export type ApproveNativeCurrencyProps = {
-  signer: Signer
-  destinationChainProvider: Provider
-}
