--- packages/arb-token-bridge-ui/src/token-bridge-sdk/core/approveToken.ts
@@ -1,23 +1,0 @@
-import { Erc20Bridger } from '@arbitrum/sdk'
-import { Provider } from '@ethersproject/providers'
-import { Signer } from 'ethers'
-import { SelectedToken } from '../BridgeTransferStarter'
-
-export type ApproveTokenProps = {
-  signer: Signer
-  selectedToken: SelectedToken
-  destinationChainProvider: Provider
-}
-
-export const approveToken = async ({
-  signer,
-  selectedToken,
-  destinationChainProvider
-}: ApproveTokenProps) => {
-  const erc20Bridger = await Erc20Bridger.fromProvider(destinationChainProvider)
-  const tx = await erc20Bridger.approveToken({
-    erc20L1Address: selectedToken.address,
-    l1Signer: signer
-  })
-  await tx.wait()
-}
