--- packages/arb-token-bridge-ui/src/token-bridge-sdk/core/getAddressFromSigner.ts
@@ -1,6 +1,0 @@
-import { Signer } from 'ethers'
-
-export const getAddressFromSigner = async (signer: Signer) => {
-  const address = await signer.getAddress()
-  return address
-}
