--- packages/arb-token-bridge-ui/src/token-bridge-sdk/core/getProviderFromSigner.ts
@@ -1,6 +1,0 @@
-import { Signer } from 'ethers'
-
-export const getProviderFromSigner = (signer: Signer) => {
-  if (!signer.provider) throw Error('Signer not able to return provider')
-  return signer.provider
-}
