+++ packages/arb-token-bridge-ui/src/token-bridge-sdk/utils.ts
@@ -1,4 +1,10 @@
+import { Signer } from 'ethers'
 import { Provider } from '@ethersproject/providers'
+
+export const getAddressFromSigner = async (signer: Signer) => {
+  const address = await signer.getAddress()
+  return address
+}
 
 export const getChainIdFromProvider = async (provider: Provider) => {
   const network = await provider.getNetwork()
